Microservice Architecture and System Design with Python & Kubernetes/Sys Design Kuberneteus/python/src/auth/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
PyJWT==2.6.0
Werkzeug==2.2.3
